Get setting from appropriate scope
Configuration scope w/ multiroot extensions is difficult. While the API
makes it easy to resolve *a* value, it doesn't make it super easy to
understand where that setting came from.

For us, in the multi-root case, we want to prompt for a folder *if* they
have a per-folder setting. Otherwise, just use whatever is resolved.
This is because two folders might configure conflicting values, and the
command be executed with no active editors.

While we could have restricted it to having an active editor, that
seemed a little restrictive.
So, we concocted the code we have here.

Additionally, the command showing in the command palette was restricted
to when there is a folder or workspace open.