Refactor task monitor to take a resolver not an array
As we get closer to monitoring tasks in the extension, it's clear that
having a fixed array at construction time is not helpful.

This switches it to a resolver (e.g. callback).

Addtionally, `getConfigurationScopeFromActiveEditor` was tweaked to use
'undefined' for the workspace, not a specific workspacefolder for
consistenty with the API.