Expose `onDidMatchingTaskExecute` from `TaskMonitor`
There is now an event exposed from `TaskMonitor` that is raised
everytime a **matching** task starts. With it, the number of times its
been raised is included -- this is to facilitate future onetime/
everytime behaviour.

With this change the promise is removed as it doesn't fit the multiple
times scenario.

Tests are updated too, 'cause I'm a good human.