Refactor task matching to return a scope
Due to the way we want to open browsers in multiroot contexts, we need
to switch from 'there was a task running. somewhere', to 'there is a
task and it's from this workspace'.

This updates the types, and introduces a new interface
`MatchedExecutionOccured` to represent this.

In a future change, the functions will be renamed for more clarity