Add delay option for auto-opened browsers + tweaked text
Some servers take a moment to startup, even after the task says it
started. Rather than have the browser open to a service that errors,
delay by a fixed amount in MS.

Also cleanred up the configuration declarations, which were all wrong
due to a missunderstanding of how they worked.