Resolve issue resolving config w/ multiroot workspaces
Tasks & Configuration can be set on multiple levels: User (global),
Workspace, Folder. At each of these levels, you can also configure settings
for this extension -- but those might be divorced from the tasks themselves.

When the task itself executes, it's scope reflects it's source. When the task
is in a folder, this works out nicely because configuration in VS Code will
look to it's parent 'container'. However, this doesn't apply in reverse. A
task thats defined at the user level, in a multiroot workspace, you don't kno
where to resolve configuration from -- folder? Which one? In the rare
workspace-but-single-folder scenario, you can just pick the single
Workspace folder.

But otherwise we're kinda SOL. We can guess by using the active editor, which
if it's not got configuration will still roll up to the parent level of