Fix test failures due to platform change (aka stop using `node-fetch`)
Due to https://github.com/microsoft/vscode/issues/130367, node-fetch
can't seem to be used in test code (And maybe extension code), causing
a failure when initial text execution happens.

Since our usage of `node-fetch` is very limited, remove it and use the
native `http` module to do the request. We wrapped it to look the same
in an effort to minimize the changes.

Follow up change will remove it from the package dependendices