Automatically append parameters to URLs in the simple browser to bypass the browser cache
When opening a browser in response to a task being executed, there is a
potential for a delay between the **task** saying it's executed, and
when that actual process starts accepting connections. We already allow
for mitigation here using the `autoDelay` setting, but that still isn't
reliable enough.

With this change, we now have an optional (on by default) setting that
will open a socket to the target host to see if it's accepting
connections on the target port; once it *is* available, we'll open the
browser. If the target isn't available before the timeout, we give up.