Fix socket checking failing on most node hosted projects
Problem:
When opening the browser, we check the target host + port to see if it's open to
ensure that it's likely going to be successful. This requires us to open a
socket to that host + port, and handle the socket connection state.

Some point in the last year, this started failing for the sample projects and
'real world' projects. It was unclear why. After some investigation, it became
clear that due to a change in the default transport preference in node there was
a mismatch between what environment we're running in, and the environment that
the local dev server is running in. Specifically node switched to prefer IPv6
hosts over IPv4, but the environment that the extension is in still prefers
IPv4.

Fix:
We explicitly test IPv4, **and** IPv6 sockets. We test IPv4 and only if that
fails do we then try IPv6. This continues to be in the time-limited loop.